+++ lib/page/homePage.dart
@@ -32,7 +32,13 @@ class _HomePageState extends State<HomePage> {
 
   NotificationPlugin _notificationPlugin;
 
-  List<String> _tabList = [];
+  List<String> _tabList = [
+    'Assets',
+    'Staking',
+    'Governance',
+    'Profile',
+  ];
+  int _tabIndex = 0;
 
   final List<String> _tabListKusama = [
     'Assets',
@@ -114,14 +120,6 @@ class _HomePageState extends State<HomePage> {
                 ),
               ],
             ),
-            bottomNavigationBar: BottomNavigationBar(
-                currentIndex: i,
-                iconSize: 22.0,
-                onTap: (index) {
-                  _pageController.jumpToPage(index);
-                },
-                type: BottomNavigationBarType.fixed,
-                items: _navBarItems(i)),
             body: _getPage(i),
           ),
         );
@@ -131,14 +129,6 @@ class _HomePageState extends State<HomePage> {
         AssetImage("assets/images/staking/top_bg_$imageColor.png"),
         Scaffold(
           backgroundColor: Colors.transparent,
-          bottomNavigationBar: BottomNavigationBar(
-              currentIndex: i,
-              iconSize: 22.0,
-              onTap: (index) {
-                _pageController.jumpToPage(index);
-              },
-              type: BottomNavigationBarType.fixed,
-              items: _navBarItems(i)),
           body: _getPage(i),
         ),
       );
@@ -177,9 +167,34 @@ class _HomePageState extends State<HomePage> {
 
   @override
   Widget build(BuildContext context) {
-    return PageView(
-      controller: _pageController,
-      children: _buildPages(),
+    return SafeArea(
+      child: Scaffold(
+        body: PageView(
+          controller: _pageController,
+          onPageChanged: (index) {
+            setState(() {
+              _tabIndex = index;
+            });
+          },
+          children: _buildPages(),
+        ),
+        bottomNavigationBar: BottomNavigationBar(
+          currentIndex: _tabIndex,
+          iconSize: 22.0,
+          onTap: (index) {
+            setState(() {
+              _tabIndex = index;
+            });
+            _pageController.animateToPage(
+              index,
+              duration: Duration(milliseconds: 500),
+              curve: Curves.ease,
+            );
+          },
+          type: BottomNavigationBarType.fixed,
+          items: _navBarItems(_tabIndex),
+        ),
+      ),
     );
   }
 }